FindBugs, Round 1, more to come FlashGot: fixes #169 Request,Cookies: moved parseCookies to Cookies Auth: removed, no longer needed
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@7245 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/captcha/pixelgrid/PixelGrid.java
@@ -1640,8 +1640,8 @@ public class PixelGrid extends Property {
      *            Grobheit
      */
     public void sampleDown(int faktor) {
-        int newWidth = (int) Math.ceil(getWidth() / faktor);
-        int newHeight = (int) Math.ceil(getHeight() / faktor);
+        int newWidth = (int) Math.ceil(getWidth() / (double)faktor);
+        int newHeight = (int) Math.ceil(getHeight() / (double)faktor);
 
         int[][] newGrid = new int[getWidth()][getHeight()];
